Clean up off_t usage
* Add comments to C++ and Fortran API calls that use off_t
* Remove noise casts for small integers

+++ test/fillval.c
@@ -2098,7 +2098,7 @@ test_extend(hid_t fapl, const char *base_name, H5D_layout_t layout)
         if ((fd = HDopen(FILE_NAME_RAW, O_RDWR | O_CREAT | O_TRUNC, H5_POSIX_CREATE_MODE_RW)) < 0 ||
             HDclose(fd) < 0)
             goto error;
-        if (H5Pset_external(dcpl, FILE_NAME_RAW, (off_t)0, (hsize_t)nelmts * sizeof(int)) < 0)
+        if (H5Pset_external(dcpl, FILE_NAME_RAW, 0, (hsize_t)nelmts * sizeof(int)) < 0)
             goto error;
     }
 #endif
